* master:
  Fix potential bugs found by Coverity analysis
  Lua test-suite can now be run out of source
  Fix out of source test-suite runs for Octave
  Add runtime test for commit 7a96fba836558f592dc5b37b32d1845f606e2efe
  Add C++11 constexpr runtime test
  Remove some useless code
  Fixes for C enums used in an API and the definition of the enum has not been parsed.
  Correct messages displayed when running Java test-suite
  Java enums - remove redundant typemaps

+++ Source/Swig/scanner.c
@@ -1557,7 +1557,7 @@ String *Scanner_get_raw_text_balanced(Scanner *s, int startchar, int endchar) {
   char c;
   int old_line = s->line;
   String *old_text = Copy(s->text);
-  int position = Tell(s->str);
+  long position = Tell(s->str);
 
   int num_levels = 1;
   int state = 0;
